refactor: extract string/multilang utilities to util_string.ts
